printer/requirements.txt
redis==6.0.0
opentracing==2.4.0
jaeger-client==4.8.0
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1